[22-10-16] 내 정보 수정 버그 해결

+++ src/main/java/com/havit/finalbe/service/MyPageService.java
@@ -108,8 +108,10 @@ public class MyPageService {
             imageId = findMember.getImageId();
         }
 
-        if (!findMember.getImageId().equals(imageId)) {
-            imageService.deleteImage(findMember.getImageId());
+        if (null != findMember.getImageId()) {
+            if (!findMember.getImageId().equals(imageId)) {
+                imageService.deleteImage(findMember.getImageId());
+            }
         }
 
         String nickname = myPageDto.getNickname();
